chore: move v2-to-v3 remove utils to utils/remove
