+++ src/transforms/v2-to-v3/utils/index.ts
@@ -1,8 +1,5 @@
 export * from "./containsRequire";
 export * from "./isTypeScriptFile";
-export * from "./removeDefaultModuleIfNotUsed";
-export * from "./removePromiseCalls";
-export * from "./removeV2ClientModule";
 export * from "./replaceClientCreation";
 export * from "./replaceTSTypeReference";
 export * from "./types";
